test: split project onboarding journey
